Декомпозиция
