+++ app/bucket/model/types.ts
